requirements.txt
django==1.9.1
djangorestframework==3.3.2
django-import-export==0.4.5
